Add shanten segment generation
Shanten for suits requires us to find runs, sets, and quads while honors just
need sets and quads. To facilitate the reduction of search space we can take an
individual suit and find segments that are of interest to us. If the calculation
is complex we can also look for traits like Edges that invalidate most valid
hands.